Updated to using Phinx to manage database migrations
