+++ src/Documentation/Searching/PostgreSqlSearchIndex.php
@@ -12,13 +12,7 @@ declare(strict_types=1);
 
 namespace App\Documentation\Searching;
 
-use DOMDocument;
-use DOMElement;
-use DOMNode;
-use Exception;
-use League\Flysystem\FilesystemOperator;
 use PDO;
-use Throwable;
 
 /**
  * Defines the documentation search index backed by PostgreSQL
@@ -27,71 +21,11 @@ final class PostgreSqlSearchIndex implements ISearchIndex
 {
     /** @var int The maximum number of search results we'll return */
     private const int MAX_NUM_SEARCH_RESULTS = 5;
-    /** @var string The name of the table */
-    private const string TABLE_NAME = 'lexemes';
-    /** @var array<string, string> The mapping of HTML elements to their weights (per PostgreSQL's setweight() function) */
-    private static array $htmlElementsToWeights = [
-        'h1' => 'A',
-        'h2' => 'B',
-        'h3' => 'C',
-        'h4' => 'D',
-        'h5' => 'D',
-        'p' => 'D',
-        'li' => 'D',
-        'blockquote' => 'D'
-    ];
 
     /**
      * @param PDO $pdo The DB connection to use
-     * @param string $linkPrefix The prefix to use for all links that are generated
-     * @param FilesystemOperator $files The file system helper
      */
-    public function __construct(
-        private readonly PDO $pdo,
-        private readonly string $linkPrefix,
-        private readonly FilesystemOperator $files
-    ) {}
-
-    /**
-     * Builds up a search index for a list of document paths
-     *
-     * @param list<string> $htmlPaths The paths to the HTML docs
-     * @throws IndexingFailedException Thrown when there was a failure to index the documents
-     */
-    public function buildSearchIndex(array $htmlPaths): void
-    {
-        if (empty($htmlPaths)) {
-            throw new IndexingFailedException('There are no documents to index');
-        }
-
-        try {
-            $indexEntries = [];
-            $dom = new DOMDocument();
-
-            foreach ($htmlPaths as $htmlPath) {
-                \libxml_use_internal_errors(true);
-                $html = $this->files->read($htmlPath);
-
-                if (empty($html) || $dom->loadHTML($html) === false) {
-                    throw new Exception("Failed to load HTML for $htmlPath: " . \strip_tags(\libxml_get_last_error()->message));
-                }
-
-                \libxml_clear_errors();
-                $h1 = $h2 = $h3 = $h4 = $h5 = null;
-
-                // Scan the documentation and index the elements
-                $body = $dom->getElementsByTagName('body')->item(0);
-
-                if ($body !== null) {
-                    $this->processNode($body, $indexEntries, $htmlPath, $h1, $h2, $h3, $h4, $h5);
-                }
-            }
-
-            $this->createAndSeedTable($indexEntries);
-        } catch (Throwable $ex) {
-            throw new IndexingFailedException('Failed to index document: ' . $ex->getMessage(), 0, $ex);
-        }
-    }
+    public function __construct(private readonly PDO $pdo) {}
 
     /**
      * @inheritdoc
@@ -109,21 +43,20 @@ final class PostgreSqlSearchIndex implements ISearchIndex
          * The goofy array_to_string(string_to_array(...)) stuff is splitting a query by spaces, adding ':*' after
          * each term, and joining them with '&' so that it forms a valid tsquery.
          */
-        $tableName = self::TABLE_NAME;
         $tsHeadlineOptions = 'StartSel=<em>, StopSel=</em>';
         $statement = $this->pdo->prepare(
             <<<EOF
 SELECT DISTINCT ON(rank, link, html_element_type) link, html_element_type, rank, h1_highlights, h2_highlights, h3_highlights, h4_highlights, h5_highlights, inner_text_highlights FROM
 ((SELECT link, html_element_type, rank, ts_headline('english', h1_inner_text, query, '{$tsHeadlineOptions}') as h1_highlights, ts_headline('english', h2_inner_text, query, '{$tsHeadlineOptions}') as h2_highlights, ts_headline('english', h3_inner_text, query, '{$tsHeadlineOptions}') as h3_highlights, ts_headline('english', h4_inner_text, query, '{$tsHeadlineOptions}') as h4_highlights, ts_headline('english', h5_inner_text, query, '{$tsHeadlineOptions}') as h5_highlights, ts_headline('english', inner_text, query, '{$tsHeadlineOptions}') as inner_text_highlights
 FROM (SELECT link, html_element_type, h1_inner_text, h2_inner_text, h3_inner_text, h4_inner_text, h5_inner_text, inner_text, ts_rank_cd(english_lexemes, query) AS rank, query
-        FROM {$tableName}, plainto_tsquery('english', :query) AS query
+        FROM lexemes, plainto_tsquery('english', :query) AS query
         WHERE english_lexemes @@ query AND (context = :context OR context = 'global')
         ORDER BY rank DESC
         LIMIT :maxResults) AS english_matching_query)
 UNION
 (SELECT link, html_element_type, rank, ts_headline(h1_inner_text, query, '{$tsHeadlineOptions}') as h1_highlights, ts_headline(h2_inner_text, query, '{$tsHeadlineOptions}') as h2_highlights, ts_headline(h3_inner_text, query, '{$tsHeadlineOptions}') as h3_highlights, ts_headline(h4_inner_text, query, '{$tsHeadlineOptions}') as h4_highlights, ts_headline(h5_inner_text, query, '{$tsHeadlineOptions}') as h5_highlights, ts_headline(inner_text, query, '{$tsHeadlineOptions}') as inner_text_highlights
     FROM (SELECT link, html_element_type, h1_inner_text, h2_inner_text, h3_inner_text, h4_inner_text, h5_inner_text, inner_text, ts_rank_cd(simple_lexemes, query) AS rank, query
-          FROM {$tableName}, (SELECT (SELECT (array_to_string(string_to_array(:query, ' '), ':* & ') || ':*'))::tsquery AS query) AS query
+          FROM lexemes, (SELECT (SELECT (array_to_string(string_to_array(:query, ' '), ':* & ') || ':*'))::tsquery AS query) AS query
           WHERE simple_lexemes @@ query AND (context = :context OR context = 'global')
           ORDER BY rank DESC
           LIMIT :maxResults) AS non_english_matching_query)) AS distinct_query
@@ -154,333 +87,5 @@ EOF
         }
 
         return $searchResults;
-    }
-
-    /**
-     * Creates and seeds the doc table
-     *
-     * @param list<IndexEntry> $indexEntries The index entries to save
-     */
-    private function createAndSeedTable(array $indexEntries): void
-    {
-        /**
-         * Update the current lexeme table name (limited to 8 chars so we don't go over PostgreSQL name length limits).
-         * This is done so that reindexing the website only impacts this instance of the site.
-         */
-        $this->pdo->beginTransaction();
-        $this->dropTable();
-        $this->createTable();
-
-        foreach ($indexEntries as $indexEntry) {
-            $this->insertIndexEntry($indexEntry);
-        }
-
-        $this->updateLexemes();
-        $this->createTableIndices();
-        $this->pdo->commit();
-    }
-
-    /**
-     * Creates an index entry
-     *
-     * @param string $filename The filename of the doc being indexed
-     * @param DOMNode $currNode The current node
-     * @param Context $context The context the entry is in
-     * @param DOMNode $h1 The current H1 node
-     * @param DOMNode|null $h2 The current H2 node
-     * @param DOMNode|null $h3 The current H3 node
-     * @param DOMNode|null $h4 The current H4 node
-     * @param DOMNode|null $h5 The current H5 node
-     * @return IndexEntry The index entry
-     * @psalm-suppress NullReference This is due to the DOM stub issues - https://github.com/vimeo/psalm/issues/5665
-     */
-    private function createIndexEntry(
-        string $filename,
-        DOMNode $currNode,
-        Context $context,
-        DOMNode $h1,
-        ?DOMNode $h2,
-        ?DOMNode $h3,
-        ?DOMNode $h4,
-        ?DOMNode $h5
-    ): IndexEntry {
-        $link = $this->linkPrefix;
-
-        /**
-         * If the current node is the h1 tag, then just link to the doc itself, not a section
-         * If the current node is a h2, h3, h4, or h5 tag, then link to the tag's ID
-         * Otherwise, find the nearest non-null header tag and link to its ID
-         */
-        if ($currNode->nodeName === 'h1') {
-            $link .= $filename;
-        } elseif (\in_array($currNode->nodeName, ['h2', 'h3', 'h4', 'h5'])) {
-            $link .= "$filename#{$currNode->attributes->getNamedItem('id')?->nodeValue}";
-        } elseif ($h5 !== null) {
-            $link .= "$filename#{$h5->attributes->getNamedItem('id')?->nodeValue}";
-        } elseif ($h4 !== null) {
-            $link .= "$filename#{$h4->attributes->getNamedItem('id')?->nodeValue}";
-        } elseif ($h3 !== null) {
-            $link .= "$filename#{$h3->attributes->getNamedItem('id')?->nodeValue}";
-        } elseif ($h2 !== null) {
-            $link .= "$filename#{$h2->attributes->getNamedItem('id')?->nodeValue}";
-        } else {
-            // h1 will never be null
-            $link .= "$filename#{$h1->attributes->getNamedItem('id')?->nodeValue}";
-        }
-
-        return new IndexEntry(
-            $currNode->nodeName,
-            $this->getAllChildNodeTexts($currNode),
-            $link,
-            self::$htmlElementsToWeights[$currNode->nodeName],
-            $context,
-            $this->getAllChildNodeTexts($h1),
-            $h2 === null ? null : $this->getAllChildNodeTexts($h2),
-            $h3 === null ? null : $this->getAllChildNodeTexts($h3),
-            $h4 === null ? null : $this->getAllChildNodeTexts($h4),
-            $h5 === null ? null : $this->getAllChildNodeTexts($h5)
-        );
-    }
-
-    /**
-     * Creates the table that will hold our docs
-     */
-    private function createTable(): void
-    {
-        $tableName = self::TABLE_NAME;
-        $statement = $this->pdo->prepare(
-            <<<EOF
-CREATE TABLE IF NOT EXISTS {$tableName} (
-    id serial primary key,
-    h1_inner_text TEXT,
-    h2_inner_text TEXT,
-    h3_inner_text TEXT,
-    h4_inner_text TEXT,
-    h5_inner_text TEXT,
-    context TEXT NOT NULL,
-    link TEXT NOT NULL,
-    html_element_type TEXT NOT NULL,
-    inner_text TEXT NOT NULL,
-    html_element_weight CHAR NOT NULL,
-    english_lexemes tsvector,
-    simple_lexemes tsvector
-)
-EOF
-        );
-        $statement->execute();
-    }
-
-    /**
-     * Creates the indices on the table for faster querying
-     */
-    private function createTableIndices(): void
-    {
-        $tableName = self::TABLE_NAME;
-        $statement = $this->pdo->prepare(
-            <<<EOF
-CREATE INDEX {$tableName}_english_lexeme_idx ON {$tableName} USING gin(english_lexemes)
-EOF
-        );
-        $statement->execute();
-        $statement = $this->pdo->prepare(
-            <<<EOF
-CREATE INDEX {$tableName}_simple_lexeme_idx ON {$tableName} USING gin(simple_lexemes)
-EOF
-        );
-        $statement->execute();
-        $statement = $this->pdo->prepare(
-            <<<EOF
-CREATE INDEX {$tableName}_context ON {$tableName}(context)
-EOF
-        );
-        $statement->execute();
-    }
-
-    /**
-     * Drops the table that hold our docs
-     */
-    private function dropTable(): void
-    {
-        $tableName = self::TABLE_NAME;
-        $statement = $this->pdo->prepare(
-            <<<EOF
-DROP TABLE IF EXISTS {$tableName}
-EOF
-        );
-        $statement->execute();
-    }
-
-    /**
-     * Recursively searches a node for all of its children's texts
-     *
-     * @param DOMNode $node The node to search
-     * @return string The children nodes' texts
-     */
-    private function getAllChildNodeTexts(DOMNode $node): string
-    {
-        $text = '';
-
-        foreach ($node->childNodes as $childNode) {
-            if ($childNode->nodeType === \XML_TEXT_NODE) {
-                $text .= $childNode->textContent;
-            } else {
-                $text .= $this->getAllChildNodeTexts($childNode);
-            }
-        }
-
-        return $text;
-    }
-
-    /**
-     * Gets the context for a node
-     *
-     * @param DOMNode $node The node whose context we want
-     * @return Context The current context
-     */
-    private function getContext(DOMNode $node): Context
-    {
-        while ($node !== null) {
-            if ($node instanceof DOMElement && $node->hasAttribute('class')) {
-                $classes = \explode(' ', $node->getAttribute('class'));
-
-                if (\in_array('context-framework', $classes, true)) {
-                    return Context::Framework;
-                }
-
-                if (\in_array('context-library', $classes, true)) {
-                    return Context::Library;
-                }
-            }
-
-            $node = $node->parentElement;
-        }
-
-        return Context::Global;
-    }
-
-    /**
-     * Inserts an index entry into the database
-     *
-     * @param IndexEntry $indexEntry The entry to insert
-     */
-    private function insertIndexEntry(IndexEntry $indexEntry): void
-    {
-        $tableName = self::TABLE_NAME;
-        $statement = $this->pdo->prepare(
-            <<<EOF
-INSERT INTO {$tableName} (h1_inner_text, h2_inner_text, h3_inner_text, h4_inner_text, h5_inner_text, link, html_element_type, inner_text, html_element_weight, context)
-VALUES (:h1InnerText, :h2InnerText, :h3InnerText, :h4InnerText, :h5InnerText, :link, :htmlElementType, :innerText, :htmlElementWeight, :context)
-EOF
-        );
-        $statement->execute([
-            'h1InnerText' => $indexEntry->h1InnerText,
-            'h2InnerText' => $indexEntry->h2InnerText,
-            'h3InnerText' => $indexEntry->h3InnerText,
-            'h4InnerText' => $indexEntry->h4InnerText,
-            'h5InnerText' => $indexEntry->h5InnerText,
-            'link' => $indexEntry->link,
-            'htmlElementType' => $indexEntry->htmlElementType,
-            'innerText' => $indexEntry->innerText,
-            'htmlElementWeight' => $indexEntry->htmlElementWeight,
-            'context' => $indexEntry->context->value
-        ]);
-    }
-
-    /**
-     * Processes a node and its children recursively, creating index entries for valid elements
-     *
-     * @param DOMNode $node The node to process
-     * @param list<IndexEntry> $indexEntries The list of index entries to update
-     * @param string $htmlPath The path to the current HTML file
-     * @param DOMNode|null $h1 The current H1 node
-     * @param DOMNode|null $h2 The current H2 node
-     * @param DOMNode|null $h3 The current H3 node
-     * @param DOMNode|null $h4 The current H4 node
-     * @param DOMNode|null $h5 The current H5 node
-     */
-    private function processNode(
-        DOMNode $node,
-        array &$indexEntries,
-        string $htmlPath,
-        ?DOMNode &$h1,
-        ?DOMNode &$h2,
-        ?DOMNode &$h3,
-        ?DOMNode &$h4,
-        ?DOMNode &$h5
-    ): void {
-        if ($this->shouldSkipProcessingNode($node)) {
-            return;
-        }
-
-        // Update headers based on the node type
-        switch ($node->nodeName) {
-            case 'h1':
-                $h1 = $node;
-                $h2 = $h3 = $h4 = $h5 = null;
-                break;
-            case 'h2':
-                $h2 = $node;
-                $h3 = $h4 = $h5 = null;
-                break;
-            case 'h3':
-                $h3 = $node;
-                $h4 = $h5 = null;
-                break;
-            case 'h4':
-                $h4 = $node;
-                $h5 = null;
-                break;
-            case 'h5':
-                $h5 = $node;
-                break;
-        }
-
-        // Only create an index entry for valid elements
-        if (isset(self::$htmlElementsToWeights[$node->nodeName])) {
-            $filename = \basename($htmlPath);
-            $indexEntries[] = $this->createIndexEntry(
-                $filename,
-                $node,
-                $this->getContext($node),
-                $h1,
-                $h2,
-                $h3,
-                $h4,
-                $h5
-            );
-        }
-
-        // Recursively process child nodes
-        foreach ($node->childNodes as $childNode) {
-            $this->processNode($childNode, $indexEntries, $htmlPath, $h1, $h2, $h3, $h4, $h5);
-        }
-    }
-
-    /**
-     * Returns whether or not we should skip processing a node and its children for lexemes
-     *
-     * @param DOMNode $node The node to check
-     * @return bool True if we should skip processing the node for lexemes, otherwise false
-     */
-    private function shouldSkipProcessingNode(DOMNode $node): bool
-    {
-        // We do not want to index the table-of-contents nav because its contents simply reflect those in the <h*> tags
-        return $node instanceof DOMElement
-            && $node->nodeName === 'nav'
-            && \str_contains($node->getAttribute('class'), 'toc-nav');
-    }
-
-    /**
-     * Updates the lexemes in all our rows
-     */
-    private function updateLexemes(): void
-    {
-        $tableName = self::TABLE_NAME;
-        $statement = $this->pdo->prepare(
-            <<<EOF
-UPDATE {$tableName} SET english_lexemes = setweight(to_tsvector('english', COALESCE(inner_text, '')), html_element_weight::"char"), simple_lexemes = setweight(to_tsvector(COALESCE(inner_text, '')), html_element_weight::"char")
-EOF
-        );
-        $statement->execute();
     }
 }
